Move metronome sound logic to actions
Removed AudioTickListener and MetronomeSynth as these are no longer needed
