+++ static/src/actions/actionTypes.js
@@ -21,6 +21,7 @@ export const DISPLAY_MESSAGE = 'DISPLAY_MESSAGE';
 export const UPDATE_METRONOME_STATUS = 'UPDATE_METRONOME_STATUS';
 export const SET_TEMPO = 'SET_TEMPO';
 export const STARTSTOP_METRONOME = 'STARTSTOP_METRONOME';
+export const SET_PLAY_SOUND = 'SET_PLAY_SOUND';
 
 // MIDI
 export const SET_MIDI_LEARN_SOUND_ID = 'SET_MIDI_LEARN_SOUND_ID';
