+++ static/src/reducers/metronome.js
@@ -1,4 +1,5 @@
-import { UPDATE_METRONOME_STATUS, SET_TEMPO, STARTSTOP_METRONOME } from '../actions/actionTypes';
+import { UPDATE_METRONOME_STATUS, SET_TEMPO, STARTSTOP_METRONOME,
+  SET_PLAY_SOUND } from '../actions/actionTypes';
 import { DEFAULT_TEMPO } from '../constants';
 import sessions from './sessions';
 
@@ -8,6 +9,7 @@ const initialState = {
   tick: 0,
   tempo: DEFAULT_TEMPO,
   isPlaying: false,
+  playSound: false,
 };
 
 function metronome(state = initialState, action) {
@@ -22,6 +24,11 @@ function metronome(state = initialState, action) {
     case SET_TEMPO: {
       return Object.assign({}, state, {
         tempo: action.tempo,
+      });
+    }
+    case SET_PLAY_SOUND: {
+      return Object.assign({}, state, {
+        playSound: action.playSound,
       });
     }
     case STARTSTOP_METRONOME: {
